kx: fix for ReprojectFeatureIterator attribute mismatch.
Requesting features from a JDBC source limiting attributes via GeoServer
WFS `&propertyname=` (or via GSR), results in an error:

    org.geotools.data.crs.ReprojectFeatureIterator.next():
    java.lang.RuntimeException: Can handle 2 attributes only, index is 2.

Because it tries to cram a feature of length 3 into a schema of length 2.
The include attribute clause gets pushed down to the JDBC query, but the
org.geotools.data.transform.TransformFeatureIteratorWrapper adds nulls
back in to match the original featuretype schema; then
ReprojectFeatureIterator fails because the full feature schema doesn't
match the filtered schema it was created with.

Retype the TransformFeatureCollection Transformer to use the filtered
schema, in the same way that we retype the TransformFeatureCollection
SimpleFeatureTypeBuilder from the change in GEOT-7601 (#4818).